[193] Add auth.py and move some of extensions
